Cost Function, Redefined Constraints, Animation
Updated the cost function to include intrusion into objects and total runtime; each with their respective weights. Updated the intrusion cost function to return a sum of the percentage of maximum intrusion (with each obstacle) along the manipulator, averaged over the time of intrusion. Fixed Minor errors in error calculation.

Corrected the definition of constraints vs input boundaries, and added kinematic constraints to the rotation of each joint.

Created an animation showing the obstacles and the robots motion over time.

Minor fixes and clean up