Intrusion Optimizer Changed
Changed the definition for intrusion cost.

Changed the optimization method to Nelder-Mead. Added the necessary cost terms to manually implement the optimization constraints.

Changed animation parameters to better plots

Added code to show the position of the robot at various moments in time environment where obstacles can be avoided. Can also do this for the results of the PD controller.